+++ saved-sites/saved-sites-impl/src/main/java/com/duckduckgo/savedsites/impl/sync/SavedSitesSyncParser.kt
@@ -16,14 +16,12 @@
 
 package com.duckduckgo.savedsites.impl.sync
 
-import com.duckduckgo.browser.api.BrowserLifecycleObserver
 import com.duckduckgo.di.scopes.AppScope
 import com.duckduckgo.savedsites.api.SavedSitesRepository
-import com.duckduckgo.sync.api.SyncChanges
-import com.duckduckgo.sync.api.SyncCrypto
-import com.duckduckgo.sync.api.SyncParser
-import com.duckduckgo.sync.api.SyncablePlugin
-import com.duckduckgo.sync.api.SyncableType.BOOKMARKS
+import com.duckduckgo.sync.api.engine.SyncChanges
+import com.duckduckgo.sync.api.engine.SyncParser
+import com.duckduckgo.sync.api.engine.SyncablePlugin
+import com.duckduckgo.sync.api.engine.SyncableType.BOOKMARKS
 import com.squareup.anvil.annotations.ContributesBinding
 import com.squareup.anvil.annotations.ContributesMultibinding
 import javax.inject.Inject
