+++ saved-sites/saved-sites-store/src/main/java/com/duckduckgo/savedsites/store/SavedSitesEntitiesDao.kt
@@ -79,7 +79,7 @@ interface SavedSitesEntitiesDao {
     @Query("update entities set lastModified = :lastModified where entityId = :entityId")
     fun updateModified(entityId: String, lastModified: String = DatabaseDateFormatter.timestamp())
 
-    @Query("select * from entities limit 1")
+    @Query("select * from entities order by rowid desc limit 1")
     fun lastModified(): Flow<Entity>
 
     @Query("select * from entities where url = :url and entities.deleted = 0 limit 1")
