+++ sync/sync-api/src/main/java/com/duckduckgo/sync/api/engine/SyncParser.kt
@@ -14,21 +14,8 @@
  * limitations under the License.
  */
 
-package com.duckduckgo.sync.impl.di
+package com.duckduckgo.sync.api.engine
 
-import com.duckduckgo.anvil.annotations.ContributesPluginPoint
-import com.duckduckgo.di.scopes.AppScope
-import com.duckduckgo.sync.api.engine.SyncablePlugin
-
-/**
- * This is here to trigger the code generations
- * [SyncablePlugin] lives in the api module but we don't want to add
- * anvil dependencies
- * CodeGen should not be generated in public api modules.
- */
-@ContributesPluginPoint(
-    scope = AppScope::class,
-    boundType = SyncablePlugin::class,
-)
-@Suppress("unused")
-private interface UnusedSyncablePluginCodegenTrigger
+interface SyncParser {
+    fun parseChanges(since: String): SyncChanges
+}
